refactor(logos): split API from utility package
