+++ apps/api/package.json
@@ -25,6 +25,7 @@
     "@opentelemetry/sdk-trace-node": "^2.1.0",
     "@scalar/hono-api-reference": "^0.9.3",
     "@sgcarstrends/database": "workspace:*",
+    "@sgcarstrends/logos": "workspace:*",
     "@sgcarstrends/types": "workspace:*",
     "@sgcarstrends/utils": "workspace:*",
     "@sindresorhus/slugify": "^2.2.1",
