+++ apps/api/src/features/logos/index.ts
@@ -1,0 +1,1 @@
+export { logosRoutes } from "./routes";
